# public/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger General - OpenAPI 3.0
  description: |-
    Web general
    - [General](/)
  version: 1.0.11
servers:
  - url: /api
...
paths:
...
  /reset-password:
    post:
      tags:
        - authentication
      summary: Reset password account user
      description: Reset password account user
      operationId: reset-password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPassword"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ResetPassword"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Successful"
        "401":
          description: Unauthorized
        "422":
          description: Invalid input
...
components:
  schemas:
    Successful:
      type: object
      properties:
        status:
          type: string
          description: Successful.
        data:
          type: array
          items: {}
          description: Data.
        message:
          type: string
          description: Message.
...
    ResetPassword:
      required:
        - token
        - email
        - password
        - password_confirmation
      type: object
      properties:
        email:
          type: string
          format: email
          example: test@gmail.com
        token:
          type: string
          example: 4d349500ee2036186dd1afa2b3d068586d12b56c0d34ed1535912f40c82190fa
        password:
          type: string
          example: 12345678
        password_confirmation:
          type: string
          example: 12345678
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT